Refactor to MVC architecture
