+++ controllers/questionController.js
@@ -1,0 +1,5 @@
+const getAllQuestions = (req, res) => {
+  return res.status(200).json({ data: { questions: QUESTIONS } });
+};
+
+export { getAllQuestions };
