+++ index.js
@@ -1,4 +1,7 @@
 import express from "express";
+import authRoutes from "./routes/authRoutes.js";
+import questionRoutes from "./routes/questionRoutes.js";
+import submissionRoutes from "./routes/submissionRoutes.js";
 
 const app = express();
 const port = 3001;
@@ -7,146 +10,9 @@ const port = 3001;
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
-const USERS = [];
-
-const QUESTIONS = [
-  {
-    id: 1,
-    title: "Two states",
-    description: "Given an array , return the maximum of the array?",
-    testCases: [
-      {
-        input: "[1,2,3,4,5]",
-        output: "5",
-      },
-    ],
-  },
-];
-
-const SUBMISSION = [];
-
-app.post("/signup", function (req, res) {
-  try {
-    const { name = null, email = null, password = null } = req.body;
-
-    if (!(name && email && password)) {
-      throw new Error("name, email and password are required!");
-    }
-
-    // Check existing users with same email
-    const emailInUse = USERS.find(
-      (user) => user.email.toLowerCase() === email.toLowerCase()
-    );
-
-    if (emailInUse) {
-      throw new Error("User with this email already exists!");
-    }
-
-    USERS.push({
-      name,
-      email: email.toLowerCase(),
-      password: password.trim(),
-    });
-
-    return res.status(201).json({ success: "Signup successful." });
-  } catch (error) {
-    return res.status(400).json({ error: error.message });
-  }
-});
-
-app.post("/login", function (req, res) {
-  try {
-    const { email = null, password = null } = req.body;
-
-    if (!(email && password)) {
-      throw new Error("Please enter your email and password!");
-    }
-
-    const existingUser = USERS.find(
-      (user) =>
-        user.email.toLowerCase() === email.toLowerCase() &&
-        user.password === password.trim()
-    );
-
-    if (!existingUser) {
-      throw new Error("Invalid credentials!");
-    }
-
-    const token = "aasldkw039uj03iurj28wiuj9w";
-
-    delete existingUser.password;
-
-    const data = {
-      user: existingUser,
-      token,
-    };
-
-    return res.status(200).json({ success: "Login successful", data });
-  } catch (error) {
-    return res.status(401).json({ error: error.message });
-  }
-});
-
-const validateToken = (req, res, next) => {
-  const { authorization = null } = req.headers;
-
-  if (!(authorization && authorization === "aasldkw039uj03iurj28wiuj9w")) {
-    return res.status(401).json({ error: "Unauthorized request!" });
-  }
-
-  next();
-};
-
-app.get("/questions", validateToken, function (req, res) {
-  return res.status(200).json({ data: { questions: QUESTIONS } });
-});
-
-app.get("/submissions", validateToken, function (req, res) {
-  return res.status(200).json({ data: { submissions: SUBMISSION } });
-});
-
-app.post("/submissions", function (req, res) {
-  try {
-    const { question_id = null, solution = null } = req.body;
-
-    if (!(question_id && solution)) {
-      throw new Error("question_id and solution are required!");
-    }
-
-    const questionExists = QUESTIONS.find(
-      (que) => que.id === parseInt(question_id)
-    );
-    if (!questionExists) {
-      throw new Error(
-        "Question associated with the given question_id not found!"
-      );
-    }
-
-    const testsPassed = Math.random() % 2 === 0;
-
-    const submission = {
-      question_id: parseInt(question_id),
-      solution: JSON.stringify(solution),
-      accepted: testsPassed,
-    };
-
-    SUBMISSION.push(submission);
-
-    if (testsPassed) {
-      return res.status(200).json({
-        success: "Your solution has been accepted.",
-        data: { submission },
-      });
-    }
-
-    return res.status(200).json({
-      error: "Your solution has been rejected!",
-      data: { submission },
-    });
-  } catch (error) {
-    return res.status(400).json({ error: error.message });
-  }
-});
+app.use("/", authRoutes);
+app.use("/questions", questionRoutes);
+app.use("/submissions", submissionRoutes);
 
 // leaving as hard todos
 // Create a route that lets an admin add a new problem
