+++ models/Question.js
@@ -1,0 +1,13 @@
+export const QUESTIONS = [
+  {
+    id: 1,
+    title: "Two states",
+    description: "Given an array , return the maximum of the array?",
+    testCases: [
+      {
+        input: "[1,2,3,4,5]",
+        output: "5",
+      },
+    ],
+  },
+];
