+++ models/Submission.js
@@ -1,0 +1,1 @@
+export const SUBMISSIONS = [];
