+++ utils/authUtils.js
@@ -1,0 +1,7 @@
+export const validateToken = (token = null) => {
+  if (!token) {
+    return false;
+  }
+
+  return token === "aasldkw039uj03iurj28wiuj9w";
+};
